cast sleep timeout to int

+++ app/web/app.py
@@ -36,7 +36,7 @@ def index():
 
     sleep_timer_timeout_in_minutes = None
     if sleep_timer.get_timeout() is not None:
-        sleep_timer_timeout_in_minutes = sleep_timer.get_timeout() / 60
+        sleep_timer_timeout_in_minutes = int(sleep_timer.get_timeout() / 60)
 
     return render_template('index.html', volume=player.get_volume(), msg=message,
                            sleep_timer_timeout=sleep_timer_timeout_in_minutes)
